Moved field provider test mock to shared mocks directory.
